# String Interning implementation