--- a/vm/runtime.h
+++ b/vm/runtime.h
@@ -3,6 +3,8 @@
 #include <cstdint>
 #include <string>
 #include <cstring>
+#include <unordered_map>
+#include  <mutex>
 
 /// Type tag, 8 bits
 typedef uint8_t Tag;
@@ -228,7 +230,6 @@ public:
     {
         return OF_DATA + (len + 1) * sizeof(char);
     }
-
     String(std::string str);
     String(Value value);
 
@@ -247,11 +248,7 @@ public:
 
     // FIXME: temporary until string interning is implemented
     bool operator == (String that) { 
-        auto len1 = length();     
-        auto len2 = that.length();     
-        const char* dataptr1 = getDataPtr(); 
-        const char* dataptr2 = that.getDataPtr(); 
-        return len1 == len2 && strncmp(dataptr1, dataptr2, len1) == 0;
+        return val == that.val;
     }
 
     /// Get the ith character code
@@ -405,6 +402,28 @@ public:
     ImgRef(Value val);
 
     std::string getName() const;
+};
+
+class StringHasher
+{
+public:
+    size_t operator()(const std::string str) const;
+};
+
+// class StringDeepComparison
+// {
+// public:
+//     bool operator()(const Value str1, const Value str2) const;
+// };
+
+class StringPool
+{
+private:
+    std::unordered_map<std::string, Value, StringHasher> pool;
+    Value addNewString(std::string str);
+public: 
+    StringPool();
+    Value getString(std::string str);
 };
 
 /// Global virtual machine instance