I implemented string interning in zeta. In order to accomplish this I used an unordered_map, but the code is relatively modular, so changing the internal implementation of the StringPool shouldn't be too difficult. (The StringPool is the class that takes care of managing strings).

I've done the benchmarks on my machine and they seem quite good!

Before:
```
benchmarks/img_fill.pls               337 ms
benchmarks/incr_field_1m.pls          641 ms
benchmarks/fcalls_10m.zim             885 ms
benchmarks/fib29.pls                  699 ms
benchmarks/loop_cnt_100m.zim         4110 ms
benchmarks/plush_parser.zim           320 ms
benchmarks/saw_wave.pls               770 ms
benchmarks/sine_wave.pls              873 ms
--------------------------------------------
geometric mean                        766 ms
``` 

After: 
```
benchmarks/img_fill.pls               282 ms
benchmarks/incr_field_1m.pls          648 ms
benchmarks/fcalls_10m.zim             929 ms
benchmarks/fib29.pls                  658 ms
benchmarks/loop_cnt_100m.zim         4299 ms
benchmarks/plush_parser.zim           203 ms
benchmarks/saw_wave.pls               614 ms
benchmarks/sine_wave.pls              781 ms
--------------------------------------------
geometric mean                        682 ms
```

The plush_parser got a nice boost!