@krypt-n 2017-06-18
Could you add a comment with a reference for the hash function?