@tommyettinger 2017-06-18
There seems to be some amount of overhead added, maybe per-call or per-loop-iteration, since the benchmarks are mostly faster except for the cases where millions of statements are run, where it seems to be that there's a slight regression. That's kinda surprising, since I don't think `loop_cnt_100m.zim` does much with strings... I haven't looked over the code yet, but does anyone have ideas why this sometimes hinders performance instead of helping it?