@maximecb 2017-06-18
@tommyettinger loop_cnt_100m is handwritten bytecode, it only has comparisons, branching, incrementing a loop counter. It uses no strings, no objects, no function calls. It's basically a metric of how fast the interpreter loop runs.

@maaarcocr Nicely done. Makes sense that plush_parser got the biggest win since it does so much string manipulation. Have to wonder if we could speed up other string ops in a useful way.

Some comments/nitpicks:

hashFunction => you should call the function murmurHash3 or whatever version it is.

class StringHasher => you should probably make this private within the StringPool class.

nitpick: I would rename addNewString to just newString.