@Maaarcocr 2017-06-18
@tommyettinger I think the benchmarks, especially the one for the loops, can have some sort of noise due to the state of RAM and CPU of the computer. I ran the benchmark again and I got a better result than both of my previous benchmarks. 

@maximecb I will have a look to see if we can have other speed-ups in other ops. 

I will rename and move the class as soon as I have time!