@tommyettinger 2017-06-18
Whew, glad to hear that issue is probably just background noise. Like @maximecb , I also wonder if other string operations might be able to benefit from this or other optimizations.