@Maaarcocr 2017-06-19
@maximecb I have done what you asked. 

One trade-off that we could try is to cache the value of this operator: https://github.com/zetavm/zetavm/blob/master/vm/runtime.cpp#L171

This value is probably calculate multiple times for the same string. Apart from this, I'm not really sure if we can do anything else.