@maximecb 2017-06-19
Thanks Marco :)

I don't think it makes sense to cache that operator, since it only gets a pointer to the string data, which is not an expensive operation. To cache the value and retrieve it from somewhere would be just as expensive as computing it.

I was thinking we could possibly speed up string concatenation. Right now, string concat is going to allocate a new string object, copy over the new strings, then look that up into the hash table to look for a match. Possibly, we could use compute the hash value of the concatenated string and do a lookup without allocating the intermediate string object.

Before getting into that though, it would make sense to check how much string concatenation we're actually doing. Another first step we could take is to avoid allocating an std::string object to do the concatenation, or to use alloca if it makes sense to do that.